Allow `custom_match_function` to be asynchronous
I'd like to make a script that accurately counts streaks for Korea,
which the data from Nomatim doesn't easily allow you to do. For that I'd
like to query a different reverse lookup service in the `custom_match_function`.

By awaiting the result of `custom_match_function`, there is no behaviour change
for existing synchronous implementations, but you can now provide a
custom match function that does an HTTP request.